requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
argostranslate==1.9.6
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ctranslate2==4.6.2
Cython==3.2.2
filelock==3.20.0
fire==0.7.1
fonttools==4.61.0
fsspec==2025.12.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
packaging==25.0
paddleocr==2.10.0
pillow==12.0.0
protobuf==6.33.2
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
python-docx==1.2.0
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
sacremoses==0.0.53
scikit-image==0.25.2
scipy==1.16.3
sentencepiece==0.2.0
shapely==2.1.2
simsimd==6.5.3
six==1.17.0
soupsieve==2.8
spidev==3.5
pyserial==3.5
stanza==1.1.1
stringzilla==4.4.0
sympy==1.14.0
termcolor==3.2.0
tifffile==2025.10.16
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
